// src/app/store/clients/clients.types.ts
export interface IClientsState {
  clients: Array<IClient>;
}

export enum E_CLIENT_KEYS {
  PASSPORT_NUMBER = "PASSPORT_NUMBER",
  FIRST_NAME = "FIRST_NAME",
  LAST_NAME = "LAST_NAME",
  BIRTH_DATE = "BIRTH_DATE",
  PERMANENT_RESIDENCE = "PERMANENT_RESIDENCE",
  TEMPORARY_RESIDENCE = "TEMPORARY_RESIDENCE",
  PHONE_NUMBER = "PHONE_NUMBER",
  EMAIL = "EMAIL",
  REGISTRATION_DATE = "REGISTRATION_DATE",
}

export interface IClient {
  [E_CLIENT_KEYS.PASSPORT_NUMBER]: string;
  [E_CLIENT_KEYS.FIRST_NAME]: string;
  [E_CLIENT_KEYS.LAST_NAME]: string;
  [E_CLIENT_KEYS.BIRTH_DATE]: string;
  [E_CLIENT_KEYS.PERMANENT_RESIDENCE]: string;
  [E_CLIENT_KEYS.TEMPORARY_RESIDENCE]: string;
  [E_CLIENT_KEYS.PHONE_NUMBER]: string;
  [E_CLIENT_KEYS.EMAIL]: string;
  [E_CLIENT_KEYS.REGISTRATION_DATE]: Date;
}